requirements/extra_apps.txt
django-haystack > 2.0
